moved fetch_single_column_value_list_for_bigquery_source_config
