+++ tests/unit_test/semantic_scholar/semantic_scholar_pipeline_test.py
@@ -18,7 +18,6 @@ from data_pipeline.semantic_scholar import (
 from data_pipeline.semantic_scholar.semantic_scholar_pipeline import (
     fetch_article_by_doi,
     fetch_article_data_from_semantic_scholar_and_load_into_bigquery,
-    fetch_single_column_value_list_for_bigquery_source_config,
     iter_article_data,
     iter_doi_for_matrix_config
 )
@@ -75,15 +74,6 @@ CONFIG_1 = SemanticScholarConfig(
 )
 
 
-@pytest.fixture(name='get_single_column_value_list_from_bq_query_mock', autouse=True)
-def _get_single_column_value_list_from_bq_query_mock():
-    with patch.object(
-        semantic_scholar_pipeline_module,
-        'get_single_column_value_list_from_bq_query'
-    ) as mock:
-        yield mock
-
-
 @pytest.fixture(name='load_given_json_list_data_from_tempdir_to_bq_mock', autouse=True)
 def _load_given_json_list_data_from_tempdir_to_bq_mock():
     with patch.object(
@@ -93,7 +83,7 @@ def _load_given_json_list_data_from_tempdir_to_bq_mock():
         yield mock
 
 
-@pytest.fixture(name='fetch_single_column_value_list_for_bigquery_source_config_mock')
+@pytest.fixture(name='fetch_single_column_value_list_for_bigquery_source_config_mock', autouse=True)
 def _fetch_single_column_value_list_for_bigquery_source_config_mock():
     with patch.object(
         semantic_scholar_pipeline_module,
@@ -112,28 +102,6 @@ def _iter_doi_for_matrix_config_mock():
 def _iter_article_data_mock():
     with patch.object(semantic_scholar_pipeline_module, 'iter_article_data') as mock:
         yield mock
-
-
-class TestFetchSingleColumnValueListForBigQuerySourceConfig:
-    def test_should_call_get_single_column_value_list_from_bq_query(
-        self,
-        get_single_column_value_list_from_bq_query_mock: MagicMock
-    ):
-        fetch_single_column_value_list_for_bigquery_source_config(BIGQUERY_SOURCE_CONFIG_1)
-        get_single_column_value_list_from_bq_query_mock.assert_called_with(
-            project_name=BIGQUERY_SOURCE_CONFIG_1.project_name,
-            query=BIGQUERY_SOURCE_CONFIG_1.sql_query
-        )
-
-    def test_should_return_doi_list_from_bq_query(
-        self,
-        get_single_column_value_list_from_bq_query_mock: MagicMock
-    ):
-        get_single_column_value_list_from_bq_query_mock.return_value = DOI_LIST
-        actual_doi_list = fetch_single_column_value_list_for_bigquery_source_config(
-            BIGQUERY_SOURCE_CONFIG_1
-        )
-        assert actual_doi_list == DOI_LIST
 
 
 class TestIterDoiForMatrixConfig:
